rename to TableBlockStore
